+++ go/go.mod
@@ -1,14 +1,8 @@
-module mod
+module your-module-name
 
 go 1.23.3
 
 require (
-	github.com/gorilla/websocket v1.5.3 // indirect
-	github.com/klauspost/compress v1.17.2 // indirect
-	github.com/nats-io/nats.go v1.37.0 // indirect
-	github.com/nats-io/nkeys v0.4.7 // indirect
-	github.com/nats-io/nuid v1.0.1 // indirect
-	github.com/pebbe/zmq4 v1.2.11 // indirect
-	golang.org/x/crypto v0.18.0 // indirect
-	golang.org/x/sys v0.16.0 // indirect
+	github.com/gorilla/websocket v1.5.3
+	github.com/pebbe/zmq4 v1.2.11
 )
